Refactor: Move subcomponents to separate folders, and do not expose them for now
