+++ packages/react/src/components/table/components/HeaderRow/HeaderRow.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import styles from './Table.module.scss';
+import styles from '../../Table.module.scss';
 
 export const HeaderRow = ({ children }) => {
   return <tr className={styles.headerRow}>{children}</tr>;
