+++ packages/react/src/components/table/components/HeaderRow/index.ts
@@ -1,0 +1,1 @@
+export * from './HeaderRow';
