+++ packages/react/src/components/table/components/SortingHeaderCell/index.ts
@@ -1,0 +1,1 @@
+export * from './SortingHeaderCell';
