+++ packages/react/src/components/table/components/TableBody/index.ts
@@ -1,0 +1,1 @@
+export * from './TableBody';
