+++ packages/react/src/components/table/components/TableContainer/index.ts
@@ -1,0 +1,1 @@
+export * from './TableContainer';
