+++ packages/react/src/components/table/components/VerticalHeaderColGroup/VerticalHeaderColGroup.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import styles from './Table.module.scss';
+import styles from '../../Table.module.scss';
 
 export const VerticalHeaderColGroup = () => {
   return (
