+++ packages/react/src/components/table/components/VerticalHeaderColGroup/index.ts
@@ -1,0 +1,1 @@
+export * from './VerticalHeaderColGroup';
