+++ packages/react/src/components/table/index.ts
@@ -1,2 +1,1 @@
-export * from './Table';
 export * from './DataTable';
